@@ -8,26 +8,26 @@
     "serve": "vite preview --port 3000"
   },
   "dependencies": {
-    "@cylynx/motif": "^0.0.17",
-    "@reduxjs/toolkit": "^1.2.3",
+    "@cylynx/motif": "0.1.0",
+    "@reduxjs/toolkit": "^1.9.5",
     "attr-accept": "^2.2.2",
-    "baseui": "^9.90.0",
+    "baseui": "13.0.0",
     "helphero": "3.5.3",
     "lodash.uniqby": "^4.7.0",
-    "neo4j-driver": "4.2.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-hook-form": "^7.3.4",
-    "react-redux": "^7.2.3",
-    "styletron-engine-atomic": "^1.4.5",
-    "styletron-react": "^5.2.7",
-    "use-neo4j": "^0.3.4"
+    "neo4j-driver": "5.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.45.2",
+    "react-redux": "^8.1.2",
+    "styletron-engine-atomic": "^1.5.0",
+    "styletron-react": "^6.1.0",
+    "use-neo4j": "^0.3.13"
   },
   "devDependencies": {
-    "@types/react": "^16.9.27",
-    "@types/react-dom": "^16.9.7",
-    "@vitejs/plugin-react-refresh": "^1.3.1",
-    "vite": "2.4.3"
+    "@types/react": "^18.2.18",
+    "@types/react-dom": "^16.2.7",
+    "@vitejs/plugin-react-swc": "^3.0.0",
+    "vite": "4.4.8"
   },
   "browserslist": {
     "production": [
