@@ -12,7 +12,7 @@
   "typings": "dist/index.d.ts",
   "module": "dist/motif.es.js",
   "engines": {
-    "node": "^14 || ^16"
+    "node": ">16"
   },
   "files": [
     "dist"
@@ -97,34 +97,34 @@
     "@types/styled-components": "^5.1.4",
     "@types/styletron-engine-atomic": "^1.1.0",
     "@types/styletron-react": "^5.0.2",
-    "@types/webpack": "^4.41.26",
-    "@vitejs/plugin-react-refresh": "^1.3.3",
-    "babel-loader": "8.2.2",
-    "css-loader": "^5.2.4",
-    "cypress": "^7.2.0",
+    "@types/webpack": "^5.28.1",
+    "@vitejs/plugin-react-swc": "^3.0.0",
+    "babel-loader": "9.1.3",
+    "css-loader": "^6.8.1",
+    "cypress": "^12.17.0",
     "cypress-file-upload": "^5.0.2",
-    "cypress-react-selector": "^2.3.6",
-    "jest": "^26.6.3",
-    "jsdom-worker": "^0.2.1",
-    "node-fetch": "^2.6.1",
+    "cypress-react-selector": "^3.0.0",
+    "jest": "^29.6.2",
+    "jsdom-worker": "^0.3.0",
+    "node-fetch": "^3.3.2",
     "redux-mock-store": "^1.5.4",
-    "rimraf": "^3.0.2",
-    "rollup-plugin-typescript2": "^0.27.3",
-    "rollup-plugin-visualizer": "^5.4.1",
-    "ts-jest": "^26.5.5",
-    "ts-loader": "^8.0.15",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.2.4",
-    "vite": "2.4.3",
-    "webpack": "4.42.0",
+    "rimraf": "^5.0.0",
+    "rollup-plugin-typescript2": "^0.35.0",
+    "rollup-plugin-visualizer": "^5.9.2",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.4.4",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.6",
+    "vite": "4.4.8",
+    "webpack": "5.88.2",
     "workerloader-jest-transformer": "^0.0.2"
   },
   "peerDependencies": {
-    "baseui": "^9.112.0",
+    "baseui": "^13.0.0",
     "react": ">=16.8.0",
     "react-dom": ">=16.8.0",
-    "styletron-engine-atomic": "^1.4.7",
-    "styletron-react": "^5.2.7"
+    "styletron-engine-atomic": "^1.5.0",
+    "styletron-react": "^6.0.0"
   },
   "publishConfig": {
     "access": "public"
