@@ -50,41 +50,41 @@
     "watch:labextension": "jupyter labextension watch ."
   },
   "dependencies": {
-    "@cylynx/motif": "^0.0.14",
+    "@cylynx/motif": "0.1.0",
     "@jupyter-widgets/base": "^1.1.10 || ^2.0.0 || ^3.0.0 || ^4.0.0",
-    "@reduxjs/toolkit": "^1.5.0",
-    "baseui": "^9.90.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-hook-form": "^7.3.4",
-    "react-redux": "^7.2.3",
-    "styletron-engine-atomic": "^1.4.5",
-    "styletron-react": "^5.2.7"
+    "@reduxjs/toolkit": "^1.9.5",
+    "baseui": "13.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.45.2",
+    "react-redux": "^8.1.2",
+    "styletron-engine-atomic": "^1.5.0",
+    "styletron-react": "^6.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.0",
-    "@babel/preset-env": "^7.5.0",
+    "@babel/core": "^7.22.9",
+    "@babel/preset-env": "^7.22.9",
     "@jupyterlab/builder": "^3.0.0",
     "@phosphor/application": "^1.6.0",
-    "@phosphor/widgets": "^1.6.0",
-    "@types/jest": "^26.0.0",
-    "@types/node": "^10.11.6",
-    "@types/webpack-env": "^1.13.6",
-    "acorn": "^7.2.0",
-    "css-loader": "^3.2.0",
-    "fs-extra": "^7.0.0",
-    "jest": "^26.0.0",
-    "mkdirp": "^0.5.1",
+    "@phosphor/widgets": "^1.9.0",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^20.4.5",
+    "@types/webpack-env": "^1.18.1",
+    "acorn": "^8.10.0",
+    "css-loader": "^6.8.1",
+    "fs-extra": "^11.1.0",
+    "jest": "^29.0.0",
+    "mkdirp": "^3.0.0",
     "npm-run-all": "^4.1.3",
-    "prettier": "^2.0.5",
-    "rimraf": "^2.6.2",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^1.0.0",
-    "ts-jest": "^26.0.0",
-    "ts-loader": "^8.0.0",
-    "typescript": "^4.2.4",
-    "webpack": "^5.0.0",
-    "webpack-cli": "^4.0.0"
+    "prettier": "^3.0.0",
+    "rimraf": "^5.0.0",
+    "source-map-loader": "^4.0.0",
+    "style-loader": "^3.3.3",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.1.6",
+    "webpack": "5.88.2",
+    "webpack-cli": "^5.1.4"
   },
   "jupyterlab": {
     "extension": "lib/plugin",
