@@ -9,17 +9,17 @@
     "serve": "vite preview --port 3000"
   },
   "dependencies": {
-    "baseui": "^9.90.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "styletron-engine-atomic": "^1.4.5",
-    "styletron-react": "^5.2.7"
+    "baseui": "^13.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styletron-engine-atomic": "^1.5.0",
+    "styletron-react": "^6.1.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.27",
-    "@types/react-dom": "^16.9.7",
-    "@vitejs/plugin-react-refresh": "^1.3.1",
-    "vite": "2.4.3"
+    "@types/react": "^18.2.18",
+    "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react-swc": "^3.0.0",
+    "vite": "4.4.8"
   },
   "browserslist": {
     "production": [
