@@ -1,9 +1,11 @@
-import reactRefresh from '@vitejs/plugin-react-refresh';
 import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react-swc';
+
+const reactPlugin = react();
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [reactRefresh()],
+  plugins: [reactPlugin],
   optimizeDeps: {
     esbuildOptions: {
       keepNames: true,
