@@ -170,7 +170,7 @@ export default {
 
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
   transformIgnorePatterns: [
-    '<rootDir>/node_modules/?!(lodash-es|@cylynx/graphin)',
+    '<rootDir>/node_modules/?!(lodash-es|@antv/graphin)',
   ],
 
   // An array of regexp pattern strings that are matched against all modules before the module loader will automatically return a mock for them
