@@ -1,4 +1,4 @@
-import { IUserEdge } from '@cylynx/graphin';
+import { IUserEdge } from '@antv/graphin';
 import { mapCollectionFields } from '../data-utils/data-utils';
 
 const noValueString = 'no-values';
