@@ -1,10 +1,14 @@
-import reactRefresh from '@vitejs/plugin-react-refresh';
 import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react-swc';
+
+const reactPlugin = react();
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [reactRefresh()],
+  plugins: [reactPlugin],
   optimizeDeps: {
-    keepNames: true,
+    esbuildOptions: {
+      keepNames: true,
+    },
   },
 });
